moved files to bigchaindb-commons
