+++ .gitignore
@@ -16,6 +16,7 @@ downloads/
 eggs/
 .eggs/
 node_modules
+rethinkdb_data
 lib/
 lib64/
 parts/
