+++ bigchaindb_common/css/scss/sidebar.scss
