+++ bigchaindb_common/css/scss/spinner.scss
