+++ bigchaindb_common/css/scss/style.scss
