+++ bigchaindb_common/css/scss/variables.scss
