+++ bigchaindb_common/js/constants/api_urls.js
