+++ bigchaindb_common/js/react/actions/account_actions.js
