+++ bigchaindb_common/js/react/actions/asset_actions.js
