+++ bigchaindb_common/js/react/alt.js
