+++ bigchaindb_common/js/react/components/spinner.js
