+++ bigchaindb_common/js/react/stores/account_store.js
