+++ bigchaindb_common/js/react/stores/asset_store.js
