+++ bigchaindb_common/js/utils/fetch_api_utils.js
