+++ bigchaindb_common/js/utils/general_utils.js
