+++ bigchaindb_common/js/utils/inject_utils.js
