+++ bigchaindb_common/js/utils/requests.js
