+++ bigchaindb_common/js/utils/url_utils.js
