+++ bigchaindb_common/python/models/accounts.py
@@ -3,10 +3,11 @@ import rethinkdb as r
 import bigchaindb.crypto
 from .assets import transfer_asset
 
+table = 'examples'
+
 
 class Account:
-    def __init__(self, app_name, bigchain, name):
-        self.app_name = app_name
+    def __init__(self, bigchain, name):
         self.bigchain = bigchain
         self.name = name
         self.sk, self.vk = bigchaindb.crypto.generate_key_pair()
@@ -25,22 +26,22 @@ class Account:
 
     def save(self):
         try:
-            r.db_create(self.app_name).run(self.bigchain.conn)
+            r.db_create(table).run(self.bigchain.conn)
         except r.ReqlOpFailedError:
             pass
 
         try:
-            r.db(self.app_name).table_create('accounts').run(self.bigchain.conn)
+            r.db(table).table_create('accounts').run(self.bigchain.conn)
         except r.ReqlOpFailedError:
             pass
 
-        user_exists = list(r.db(self.app_name)
+        user_exists = list(r.db(table)
                            .table('accounts')
                            .filter(lambda user: user['name'] == self.name)
                            .run(self.bigchain.conn))
 
         if not len(user_exists):
-            r.db(self.app_name)\
+            r.db(table)\
                 .table('accounts')\
                 .insert(self.as_dict(), durability='hard')\
                 .run(self.bigchain.conn)
@@ -57,7 +58,7 @@ class Account:
         }
 
 
-def retrieve_accounts(app_name, bigchain):
-    return list(r.db(app_name)
+def retrieve_accounts(bigchain):
+    return list(r.db(table)
                 .table('accounts')
                 .run(bigchain.conn))
