+++ bigchaindb_common/python/models/assets.py
