+++ init_db.py
@@ -1,8 +1,6 @@
-import subprocess
-
 import bigchaindb
 
-from bigchaindb_common.python.accounts import Account
+from bigchaindb_common.python.models.accounts import Account
 
 app_name = 'on_the_record'
 bigchain = bigchaindb.Bigchain()
@@ -12,8 +10,7 @@ def main():
     capacity = 5
     accounts = []
     for i in range(2 * capacity):
-        account = Account(app_name=app_name,
-                          bigchain=bigchain,
+        account = Account(bigchain=bigchain,
                           name='account_{}'.format(i))
         accounts.append(account)
     print('INIT: {} accounts initialized'.format(len(accounts)))
