--- on_the_record/client/js/constants/application_constants.js
@@ -1,12 +1,0 @@
-'use strict';
-
-const apiEndpoint = process.env.NODE_ENV === 'production' ?
-    'http://ascribe-diamonds-explorer.herokuapp.com/api/' : 'http://localhost:8000/api/';
-const baseUrl = '/';
-
-const constants = {
-    baseUrl,
-    apiEndpoint
-};
-
-export default constants;
