--- on_the_record/client/scss/fonts.css
@@ -1,41 +1,0 @@
-/* This font was generated using icomoon */
-@font-face {
-    font-family: 'ascribe-icons';
-    src:url('../fonts/ascribe-icons.eot');
-    src:url('../fonts/ascribe-icons.eot') format('embedded-opentype'),
-        url('../fonts/ascribe-icons.ttf') format('truetype'),
-        url('../fonts/ascribe-icons.woff') format('woff'),
-        url('../fonts/ascribe-icons.svg') format('svg');
-    font-weight: normal;
-    font-style: normal;
-}
-
-[class^="icon-"], [class*=" icon-"] {
-    font-family: 'ascribe-icons';
-    speak: none;
-    font-style: normal;
-    font-weight: normal;
-    font-variant: normal;
-    text-transform: none;
-    line-height: 1;
-
-    /* Better Font Rendering =========== */
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-}
-
-.icon-twitter-share:before {
-    content: "\e801";
-}
-.icon-facebook-share:before {
-    content: "\e802";
-}
-.icon-email-share:before {
-    content: "\e803";
-}
-.icon-ascribe-logo-small:before {
-    content: "\e804";
-}
-.icon-ascribe-logo-normal:before {
-    content: "\e805";
-}
