+++ on_the_record/client/scss/main.scss
@@ -3,10 +3,9 @@
 @import url(https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css);
 
 $icon-font-path: "~bootstrap-sass/assets/fonts/bootstrap/";
-@import "../../node_modules/bootstrap-sass/assets/stylesheets/_bootstrap.scss";
-@import 'normalize.css';
-@import 'fonts.css';
-@import 'custom_bootstrap';
-@import 'ascribe_spinner';
-@import 'sidebar';
-@import 'style';
+@import "../../../node_modules/bootstrap-sass/assets/stylesheets/_bootstrap.scss";
+@import '../../../bigchaindb_common/css/scss/normalize.css';
+@import '../../../bigchaindb_common/css/scss/custom_bootstrap';
+@import '../../../bigchaindb_common/css/scss/spinner';
+@import '../../../bigchaindb_common/css/scss/sidebar';
+@import '../../../bigchaindb_common/css/scss/style';
